feat: ajustes de banco e de fluxo do produto

+++ src/main/kotlin/br/com/managerfood/golden/adapter/controller/dto/response/ProductResponse.kt
@@ -1,11 +1,12 @@
 package br.com.managerfood.golden.adapter.controller.dto.response
 
 import br.com.managerfood.golden.domain.Category
+import java.math.BigDecimal
 
 class ProductResponse(
     var id: Long,
     var name: String,
     var description: String,
-    var price: Float,
+    var price: BigDecimal,
     var category: MutableList<CategoryResponse> = mutableListOf()
 )
